# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hassette Web API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/telemetry/status": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "Telemetry Status",
        "description": "Health check for the telemetry database.\n\nRuns a representative query against the unified ``executions`` table.\nReturns 503 with ``degraded: true`` when the database is\nunavailable; 200 with ``degraded: false`` when healthy.\n\nA probe, so a failure answers with this status body rather than a problem body: the CLI\nand container health checks read it as data.",
        "operationId": "telemetry_status_api_telemetry_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryStatusResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryStatusResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/telemetry/app/{app_key}/listeners": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "App Listeners",
        "description": "Listener metrics with human-readable handler summaries.",
        "operationId": "app_listeners_api_telemetry_app__app_key__listeners_get",
        "parameters": [
          {
            "name": "app_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Use `__hassette__` to query framework-internal actor telemetry.",
              "title": "App Key"
            },
            "description": "Use `__hassette__` to query framework-internal actor telemetry."
          },
          {
            "name": "instance_index",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "App instance index. Defaults to 0. Multi-instance apps have indices 0..N-1.",
              "default": 0,
              "title": "Instance Index"
            },
            "description": "App instance index. Defaults to 0. Multi-instance apps have indices 0..N-1."
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "name": "source_tier",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "app",
                "framework",
                "all"
              ],
              "type": "string",
              "description": "Filter by source tier. 'app' excludes framework internals. 'framework' returns only internal actors. 'all' returns everything.",
              "default": "app",
              "title": "Source Tier"
            },
            "description": "Filter by source tier. 'app' excludes framework internals. 'framework' returns only internal actors. 'all' returns everything."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListenerWithSummary"
                  },
                  "title": "Response App Listeners Api Telemetry App  App Key  Listeners Get"
                }
              }
            }
          },
          "503": {
            "description": "`telemetry_unavailable`: the telemetry store could not be read",
            "x-problem-codes": [
              "telemetry_unavailable"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/telemetry/app/{app_key}/activity": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "App Activity",
        "description": "Recent handler invocations and job executions for a single app, merged and sorted by time.",
        "operationId": "app_activity_api_telemetry_app__app_key__activity_get",
        "parameters": [
          {
            "name": "app_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Use `__hassette__` to query framework-internal actor telemetry.",
              "title": "App Key"
            },
            "description": "Use `__hassette__` to query framework-internal actor telemetry."
          },
          {
            "name": "instance_index",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "App instance index. None returns activity across all instances.",
              "title": "Instance Index"
            },
            "description": "App instance index. None returns activity across all instances."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "name": "source_tier",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "app",
                "framework",
                "all"
              ],
              "type": "string",
              "description": "Filter by source tier. 'app' excludes framework internals. 'framework' returns only internal actors. 'all' returns everything.",
              "default": "app",
              "title": "Source Tier"
            },
            "description": "Filter by source tier. 'app' excludes framework internals. 'framework' returns only internal actors. 'all' returns everything."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityFeedEntry"
                  },
                  "title": "Response App Activity Api Telemetry App  App Key  Activity Get"
                }
              }
            }
          },
          "503": {
            "description": "`telemetry_unavailable`: the telemetry store could not be read",
            "x-problem-codes": [
              "telemetry_unavailable"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/telemetry/app/{app_key}/jobs": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "App Jobs",
        "description": "Job summaries for a single app instance, enriched with live registry data.\n\n``schedule_status``/``schedule_status_reason`` and, for ``SCHEDULED`` jobs, live timing\n(``next_run``, ``fire_at``, ``jitter``) are joined from the live scheduler registry by\n``db_id``. On registry failure the DB rows are returned without enrichment (degraded but\nfunctional; logged warning, no 500).",
        "operationId": "app_jobs_api_telemetry_app__app_key__jobs_get",
        "parameters": [
          {
            "name": "app_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Use `__hassette__` to query framework-internal actor telemetry.",
              "title": "App Key"
            },
            "description": "Use `__hassette__` to query framework-internal actor telemetry."
          },
          {
            "name": "instance_index",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "App instance index. Defaults to 0. Multi-instance apps have indices 0..N-1.",
              "default": 0,
              "title": "Instance Index"
            },
            "description": "App instance index. Defaults to 0. Multi-instance apps have indices 0..N-1."
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "name": "source_tier",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "app",
                "framework",
                "all"
              ],
              "type": "string",
              "description": "Filter by source tier. 'app' excludes framework internals. 'framework' returns only internal actors. 'all' returns everything.",
              "default": "app",
              "title": "Source Tier"
            },
            "description": "Filter by source tier. 'app' excludes framework internals. 'framework' returns only internal actors. 'all' returns everything."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobSummary"
                  },
                  "title": "Response App Jobs Api Telemetry App  App Key  Jobs Get"
                }
              }
            }
          },
          "503": {
            "description": "`telemetry_unavailable`: the telemetry store could not be read",
            "x-problem-codes": [
              "telemetry_unavailable"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/telemetry/executions": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "List Executions",
        "description": "Combined execution list (handler invocations and job executions).\n\nFilter by ``kind=handler`` or ``kind=job`` to restrict to one type.\nEach record includes a ``kind`` field that discriminates the execution type.",
        "operationId": "list_executions_api_telemetry_executions_get",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "handler",
                    "job"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by kind: 'handler' or 'job'.",
              "title": "Kind"
            },
            "description": "Filter by kind: 'handler' or 'job'."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Execution"
                  },
                  "title": "Response List Executions Api Telemetry Executions Get"
                }
              }
            }
          },
          "503": {
            "description": "`telemetry_unavailable`: the telemetry store could not be read",
            "x-problem-codes": [
              "telemetry_unavailable"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/telemetry/listener/{listener_id}/executions": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "Listener Executions",
        "description": "Execution history for a specific listener (handler invocations).",
        "operationId": "listener_executions_api_telemetry_listener__listener_id__executions_get",
        "parameters": [
          {
            "name": "listener_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Listener Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Execution"
                  },
                  "title": "Response Listener Executions Api Telemetry Listener  Listener Id  Executions Get"
                }
              }
            }
          },
          "503": {
            "description": "`telemetry_unavailable`: the telemetry store could not be read",
            "x-problem-codes": [
              "telemetry_unavailable"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/telemetry/execution/{execution_id}": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "Get Execution",
        "description": "Return a single execution record by its UUID.",
        "operationId": "get_execution_api_telemetry_execution__execution_id__get",
        "parameters": [
          {
            "name": "execution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Execution Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Execution"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Execution Api Telemetry Execution  Execution Id  Get"
                }
              }
            }
          },
          "503": {
            "description": "`telemetry_unavailable`: the telemetry store could not be read",
            "x-problem-codes": [
              "telemetry_unavailable"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TelemetryStatusResponse": {
        "properties": {
          "degraded": {
            "type": "boolean",
            "title": "Degraded"
          },
          "dropped_overflow": {
            "type": "integer",
            "title": "Dropped Overflow",
            "default": 0
          },
          "dropped_exhausted": {
            "type": "integer",
            "title": "Dropped Exhausted",
            "default": 0
          },
          "dropped_shutdown": {
            "type": "integer",
            "title": "Dropped Shutdown",
            "default": 0
          },
          "dropped_filtered": {
            "type": "integer",
            "title": "Dropped Filtered",
            "default": 0
          },
          "error_handler_failures": {
            "type": "integer",
            "title": "Error Handler Failures",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "degraded"
        ],
        "title": "TelemetryStatusResponse",
        "description": "Health check response for the telemetry database."
      }
    }
  }
}